requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
python-dotenv>=0.19.1
accelerate>=0.26.0
dataset